backend/app/services/inference_service.py
import os
import uuid
from typing import Tuple, Optional
from PIL import Image
import numpy as np

async def run_model_inference(
    sample_id: str,
    image_path: str,
    model_type: str = "mock"
) -> Tuple[bool, float, Optional[int], Optional[float], Optional[float], str, str]:
    """
    Run model inference on a rooftop image.
    
    Args:
        sample_id: The sample ID
        image_path: Path to the image
        model_type: Type of model to use (mock, yolov8, roboflow, etc.)
        
    Returns:
        Tuple of (has_solar, confidence, panel_count_est, pv_area_sqm_est, capacity_kw_est, bbox_or_mask, reason_codes)
    """
    if model_type == "roboflow":
        # Use Roboflow API for inference
        # Import here to avoid circular imports
        from .roboflow_service import run_roboflow_inference
        roboflow_api_key = os.getenv("ROBOFLOW_API_KEY")
        return await run_roboflow_inference(image_path, roboflow_api_key)
    elif model_type == "mock":
        return await _run_mock_inference(sample_id, image_path)
    else:
        # In a real implementation, we would run the actual model
        # For now, we'll just run the mock inference
        return await _run_mock_inference(sample_id, image_path)

async def _run_mock_inference(sample_id: str, image_path: str) -> Tuple[bool, float, Optional[int], Optional[float], Optional[float], str, str]:
    """
    Run mock model inference.
    
    Args:
        sample_id: The sample ID
        image_path: Path to the image
        
    Returns:
        Tuple of (has_solar, confidence, panel_count_est, pv_area_sqm_est, capacity_kw_est, bbox_or_mask, reason_codes)
    """
    # Simple mock logic based on sample_id
    site_num = int(sample_id.replace('site_', ''))
    
    # Alternate between having solar and not having solar
    has_solar = site_num % 2 == 1
    
    if has_solar:
        confidence = 0.92
        panel_count_est = 12
        pv_area_sqm_est = 20.5
        capacity_kw_est = 3.5
        bbox_or_mask = "bbox_data_mock"
        reason_codes = "module_grid,rectilinear_array"
    else:
        confidence = 0.3
        panel_count_est = None
        pv_area_sqm_est = None
        capacity_kw_est = None
        bbox_or_mask = "no_solar_detected"
        reason_codes = "no_panels_detected"
    
    return (
        has_solar,
        confidence,
        panel_count_est,
        pv_area_sqm_est,
        capacity_kw_est,
        bbox_or_mask,
        reason_codes
    )