backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
redis==5.0.1
rq==1.15.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1
pillow==10.1.0
reportlab==4.0.7
qrcode==7.4.2
minio==7.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.13.3
prometheus-client==0.19.0
python-bidi==0.4.2
googletrans==3.1.0a0
torch==2.1.1
torchvision==0.16.1
pydantic-settings==2.0.3
scikit-learn==1.3.0
numpy==1.24.3